YUK-25192 Cell LCR - Cloud data doesn't seem to be requested after Control is Active

+++ yukon-client/common/src/com/cannontech/dr/eatonCloud/EatonCloudMessageListener.java
@@ -271,7 +271,7 @@ public class EatonCloudMessageListener {
                 devices.size(),
                 eventId, 
                 command.getVirtualRelayId(), 
-                totalFailed.intValue(), 
+                totalFailed, 
                 totalSucceeded,
                 readTimeFromNowInMinutes,
                 nextReadTime.toDateTime().toString("MM-dd-yyyy HH:mm:ss.SSS"));
